initial player character setup with some placeholder assets
I added a few placeholder assets for the player, tiles, gun and bullet to try to setup basic platformer character movement and firing bullet mechanics

Known issue: Direction/Firing angle of the bullet.

Possibly the firing behaviour could be refactored so that it is used for both player and enemy assets without redoing the same setup twice or more times.